Omitted PHP closing tag
http://php.net/basic-syntax.instruction-separation#language.basic-syntax.instruction-separation
The closing tag of a PHP block at the end of a file is optional, and in
some cases omitting it is helpful when using include or require, so
unwanted whitespace will not occur at the end of files, and you will
still be able to add headers to the response later. It is also handy if
you use output buffering, and would not like to see added unwanted
whitespace at the end of the parts generated by the included files.

http://framework.zend.com/manual/1.12/en/coding-standard.php-file-formatting.html
For files that contain only PHP code, the closing tag ("?>") is never
permitted. It is not required by PHP, and omitting it prevents the
accidental injection of trailing white space into the response.